requirements.txt
numpy
pandas
opencv-python
pyqt5
PyYAML
tables
scikit-image
easydict